requirements.txt
numpy
librosa==0.8.1
scipy
tqdm
neptune-client
pyyaml
pesq
pystoi
